restructured some code
