+++ includes/minishell.h
@@ -33,6 +33,7 @@ char	*get_instructions_line(int interactive, int fd);
 void	parse(char *line);
 void	lexer(char *line);
 t_token	*get_tokens(const char *s, char c, int words);
+void	get_tokens_types(t_token *tokens, int nb_tokens);
 
 #endif
