+++ srcs/core_utils/del_memory_utils.c
