+++ srcs/parsing/lexer_utils.c
@@ -1,0 +1,11 @@
+#include "minishell.h"
+
+int		get_array_len(char **split)
+{
+	int	i;
+
+	i = 0;
+	while (split[i])
+		i++;
+	return (i);
+}
