--- srcs/parsing/parser_utils.c
@@ -1,21 +1,0 @@
-#include "minishell.h"
-
-char	*skip_space(char *line)
-{
-	int	i;
-
-	i = 0;
-	while (line[i] && line[i] == SPACE)
-		i++;; //faire la fonction
-	return (line + i);
-}
-
-int		get_array_len(char **split)
-{
-	int	i;
-
-	i = 0;
-	while (split[i])
-		i++;
-	return (i);
-}
