+++ srcs/parsing/tokens_utils.c
@@ -1,0 +1,22 @@
+#include "minishell.h"
+
+void	get_tokens_types(t_token *tokens, int nb_tokens)
+{
+	/*if (SGL_QUOTE)
+		remove_quote + WORD;
+	if (DBL_QUOTE)
+		remove_quote + deal with ENV_VAR;
+	if (1rst Token)
+		CMD;
+	else if (-something)
+		FLAG;
+	else if ($)
+		ENV_VAR;
+	else if (> < >> <<)
+		REDIRECTION;
+	else
+		WORD;*/
+	(void)tokens;
+	(void)nb_tokens;
+	return ;
+}
